+++ src/main.js
@@ -1,26 +1,26 @@
 import "core-js/stable";
 import "regenerator-runtime/runtime";
-import Vue from 'vue'
-import VueAnalytics from 'vue-analytics'
+import Vue from 'vue';
+import VueAnalytics from 'vue-analytics';
 import router from "./router";
-import App from './App.vue'
-import uswds from 'uswds'
-import browserDetect from 'vue-browser-detect-plugin'
+import App from './App.vue';
+import uswds from 'uswds';
+import browserDetect from 'vue-browser-detect-plugin';
 
-import { library } from '@fortawesome/fontawesome-svg-core'
-import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
+import { library } from '@fortawesome/fontawesome-svg-core';
+import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 
-import { faInfo } from '@fortawesome/free-solid-svg-icons'
-import { faLayerGroup } from '@fortawesome/free-solid-svg-icons'
-import { faTimes } from '@fortawesome/free-solid-svg-icons'
-import { faThumbsUp } from '@fortawesome/free-solid-svg-icons'
-import { faThumbsDown } from '@fortawesome/free-solid-svg-icons'
-import { faMinus } from '@fortawesome/free-solid-svg-icons'
-import { faPlus } from '@fortawesome/free-solid-svg-icons'
-import { faWater } from '@fortawesome/free-solid-svg-icons'
-import { faAngleDown } from '@fortawesome/free-solid-svg-icons'
-import { faAngleLeft } from '@fortawesome/free-solid-svg-icons'
-import { faList } from '@fortawesome/free-solid-svg-icons'
+import { faInfo } from '@fortawesome/free-solid-svg-icons';
+import { faLayerGroup } from '@fortawesome/free-solid-svg-icons';
+import { faTimes } from '@fortawesome/free-solid-svg-icons';
+import { faThumbsUp } from '@fortawesome/free-solid-svg-icons';
+import { faThumbsDown } from '@fortawesome/free-solid-svg-icons';
+import { faMinus } from '@fortawesome/free-solid-svg-icons';
+import { faPlus } from '@fortawesome/free-solid-svg-icons';
+import { faWater } from '@fortawesome/free-solid-svg-icons';
+import { faAngleDown } from '@fortawesome/free-solid-svg-icons';
+import { faAngleLeft } from '@fortawesome/free-solid-svg-icons';
+import { faList } from '@fortawesome/free-solid-svg-icons';
 import { faQuestion} from "@fortawesome/free-solid-svg-icons";
 import { faBackward } from "@fortawesome/free-solid-svg-icons";
 
@@ -60,11 +60,26 @@ Vue.config.productionTip = false;
 Vue.use(uswds);
 Vue.use(browserDetect);
 
+
+// create a unique ID that can be used to identify user sessions
+// Note: this section is borrowed code with Public Domain/MIT licensing
+const sessionID = function() {
+  var d = new Date().getTime();
+  if (typeof performance !== 'undefined' && typeof performance.now === 'function'){
+    d += performance.now(); //use high-precision timer if available
+  }
+  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
+    var r = (d + Math.random() * 16) % 16 | 0;
+    d = Math.floor(d / 16);
+    return (c === 'x' ? r : (r & 0x3 | 0x8)).toString(16);
+  });
+};
+
 Vue.use(VueAnalytics, {
   id: 'UA-149352326-1',
   set: [
-    { field: 'sessionId', value: 'test1' },
-    { field: 'timestamp', value: 'test2' },
+    { field: 'dimension1', value: sessionID() },
+    { field: 'dimension2', value: Date.now() },
   ],
   commands: {
     trackName (eventName, action, label) {
@@ -72,9 +87,9 @@ Vue.use(VueAnalytics, {
     }
   },
   router
-})
+});
 
 const app = new Vue({
   router,
   render: h => h(App)
-}).$mount('#app')
+}).$mount('#app');